% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{Config}
\alias{Config}
\title{Config}
\description{
A config prototype wrapper
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character string representing the name of the dataset configuration.}

\item{\code{uri}}{A character string representing the URI where the dataset comes from.}

\item{\code{ann_col}}{A character string representing the name of the annotation column. When calling \code{data@meta.data[[ann_col]]} from a Seurat object, the column with annotation should be returned.}

\item{\code{cell_types}}{A named list containing mapping of common immune cell types to annotation in the dataset.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Config-new}{\code{Config$new()}}
\item \href{#method-Config-cloneWith}{\code{Config$cloneWith()}}
\item \href{#method-Config-print}{\code{Config$print()}}
\item \href{#method-Config-clone}{\code{Config$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Config-new"></a>}}
\if{latex}{\out{\hypertarget{method-Config-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Config$new(name = NULL, uri = NULL, ann_col = "cell_type", cell_types = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string representing the name of the dataset configuration.}

\item{\code{uri}}{A character string representing the URI where the dataset comes from.}

\item{\code{ann_col}}{A character string representing the name of the annotation column. When calling \code{data@meta.data[[ann_col]]} from a Seurat object, the column with annotation should be returned.}

\item{\code{cell_types}}{A named list containing mapping of common immune cell types to annotation in the dataset.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Config-cloneWith"></a>}}
\if{latex}{\out{\hypertarget{method-Config-cloneWith}{}}}
\subsection{Method \code{cloneWith()}}{
Create a new configuration object with updated attributes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Config$cloneWith(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional parameters to update the configuration attributes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A cloned Config object with updated attributes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Config-print"></a>}}
\if{latex}{\out{\hypertarget{method-Config-print}{}}}
\subsection{Method \code{print()}}{
override print
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Config$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Config-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Config-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Config$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
