% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{Convert.curatedAtlasToSeurat}
\alias{Convert.curatedAtlasToSeurat}
\title{Convert$curatedAtlasToSeurat}
\usage{
dir <- "./data/kidney/"
Convert$curatedAtlasToSeurat(
  paste0(dir, "Exp_data_UMIcounts.mtx"),
  paste0(dir, "Cells.csv"),
  paste0(dir, "Genes.txt"),
  "type"
)
}
\arguments{
\item{mtx_path}{path to the scRNA-seq expression matrix, typically named as
\verb{*.mtx}}

\item{cell_info_path}{path to the cell information and meta data csv file,
typically named as \code{Cells.csv}}

\item{gene_name_path}{path to the gene name file, typically named as
\code{Genes.txt}}

\item{source_col}{column name in cell info csv indicating the source of the
cell, either from tumor or normal tissues}
}
\value{
Seurat data
}
\description{
Read and convert data downloaded from
\href{https://www.weizmann.ac.il/sites/3CA/}{3ca} to a Seurat object
}
