% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{Convert.ensemblToSymbol}
\alias{Convert.ensemblToSymbol}
\title{Convert$ensemblToSymbol}
\arguments{
\item{data}{Seurat data}
}
\value{
Converted data
}
\description{
Convert gene names from ensembl id to symbols
}
