% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\docType{data}
\name{Convert}
\alias{Convert}
\title{Convert}
\format{
An object of class \code{environment} of length 2.
}
\usage{
Convert
}
\description{
An object with functions that help convert between data
}
\section{Functions}{

\itemize{
\item \code{\link[=Convert.ensemblToSymbol]{Convert$ensemblToSymbol}}
\item \code{\link[=Convert.curatedAtlasToSeurat]{Convert$curatedAtlasToSeurat}}
}
}

\keyword{datasets}
