% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoader.R
\name{DataLoader.loadTissueBloodAnnotatedData}
\alias{DataLoader.loadTissueBloodAnnotatedData}
\title{DataLoader$loadTissueBloodAnnotatedData}
\arguments{
\item{config}{Dataset config}

\item{data}{Original data without tissue blood annotation. If not provided,
data will be loaded from \verb{data/$\{config$name\}.rds}.}
}
\description{
Load tissue and blood annotation into a Seurat object
}
