% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoader.R
\name{DataLoader.loadTissueBloodAnnotation}
\alias{DataLoader.loadTissueBloodAnnotation}
\title{DataLoader$loadTissueBloodAnnotation}
\arguments{
\item{config}{Dataset config}
}
\value{
Meta data as a data frame. The data frame has cell name as row name,
with two columns \code{cluster} and \code{origin_tb}. \code{cluster} is either 1 or 2, and
\code{origin_tb} is either "Blood" or "Tissue". Refer to \code{origin_tb} for the
annotation, do not use \code{cluster} to annotate blood or tissue cell.
}
\description{
Load and gather tissue and blood annotation data from Originator RData output

Originator outputs must be stored under \verb{output/$\{config$name\}/}
}
