% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoader.R
\docType{data}
\name{DataLoader}
\alias{DataLoader}
\title{DataLoader}
\format{
An object of class \code{environment} of length 2.
}
\usage{
DataLoader
}
\description{
An object with functions that help load data
}
\section{Functions}{

\itemize{
\item \code{\link[=DataLoader.loadTissueBloodAnnotation]{DataLoader$loadTissueBloodAnnotation}}
\item \code{\link[=DataLoader.loadTissueBloodAnnotatedData]{DataLoader$loadTissueBloodAnnotatedData}}
}
}

\keyword{datasets}
