% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Integrate.R
\name{Integrate.annotateRefQuery}
\alias{Integrate.annotateRefQuery}
\title{Integrate$annotateRefQuery}
\arguments{
\item{config}{Dataset config}

\item{data}{Integrated data}
}
\value{
Integrated data with ref and query cells annotated
}
\description{
Annotate ref and query identity for originator
}
