% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Integrate.R
\name{Integrate.do}
\alias{Integrate.do}
\title{Integrate$do}
\arguments{
\item{blood}{Blood reference data}

\item{query}{Query data}
}
\value{
Integrated data
}
\description{
Do integration. Use default Seurat pipeline (CCA), which is very slow. For
better options, use \code{\link[=Integrate.doHarmony]{Integrate$doHarmony}}.

\strong{This function is deprecated, do not use.}
}
