% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Integrate.R
\name{Integrate.doHarmony}
\alias{Integrate.doHarmony}
\title{Integrate$doHarmony}
\arguments{
\item{blood}{Blood reference data}

\item{query}{Query data}
}
\value{
Integrated data
}
\description{
Do integration using harmony
}
