% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Integrate.R
\docType{data}
\name{Integrate}
\alias{Integrate}
\title{Integrate}
\format{
An object of class \code{environment} of length 3.
}
\usage{
Integrate
}
\description{
An object with functions that help integrate query data with reference data
}
\section{Functions}{

\itemize{
\item \code{\link[=Integrate.do]{Integrate$do}}
\item \code{\link[=Integrate.doHarmony]{Integrate$doHarmony}}
\item \code{\link[=Integrate.annotateRefQuery]{Integrate$annotateRefQuery}}
}
}

\keyword{datasets}
