% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{integrateBloodRef}
\alias{integrateBloodRef}
\title{Use harmony to integrate query data with blood reference data}
\usage{
integrateBloodRef(blood, query, into_col = "orig.ident", seed = 42)
}
\arguments{
\item{blood}{Blood reference data}

\item{query}{Query data}

\item{into_col}{Column name to indicate if sample is from query or reference}

\item{seed}{Random seed}
}
\value{
Integrated data
}
\description{
Use harmony to integrate query data with blood reference data
}
