% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logger.R
\docType{data}
\name{Logger}
\alias{Logger}
\title{Logger}
\format{
An object of class \code{environment} of length 5.
}
\usage{
Logger
}
\description{
A logger object
}
\section{Functions}{

\itemize{
\item \code{\link[=Logger.log]{Logger$log}}
\item \code{\link[=Logger.info]{Logger$info}}
\item \code{\link[=Logger.debug]{Logger$debug}}
\item \code{\link[=Logger.warn]{Logger$warn}}
\item \code{\link[=Logger.error]{Logger$error}}
}
}

\keyword{datasets}
