% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{mapCellTypes}
\alias{mapCellTypes}
\title{Map cell type annotation}
\usage{
mapCellTypes(data, use_col, celltype_map, into_col = "unified_celltype")
}
\arguments{
\item{data}{Data}

\item{use_col}{Cell type annotation column}

\item{into_col}{Column to save mapped annotation}
}
\value{
Processed data
}
\description{
Map cell type annotation
}
