% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Originator.R
\docType{package}
\name{Originator-package}
\alias{Originator}
\alias{Originator-package}
\title{Originator}
\description{
Originator: Computational Framework Separating Single-Cell RNA-Seq by Genetic
and Contextual Information
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
