% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Originator.R
\name{originator}
\alias{originator}
\title{Perform blood and tissue-resident immune cell identification using Originator}
\usage{
originator(data, query_cell_type, plot = FALSE, output_dir = NA)
}
\arguments{
\item{data}{Seurat onject (S4 or S5) containing cell types to query. The input data is the integrated data between query and reference data. Make sure cell types are annotated as "Query \if{html}{\out{<celltype name>}}" and "Ref \if{html}{\out{<celltype name>}}" in column "annotation_query_ref"}

\item{plot}{(Default FALSE) If TRUE, generate relevant plots in the directory specify in output_dir argument}

\item{output_dir}{If plot = TRUE, output_dir needs to be specified.}

\item{query_cell_typeGiven}{"Query \if{html}{\out{<celltype name>}}" and "Ref \if{html}{\out{<celltype name>}}" in an input Seurat object, specify query_cell_type as \if{html}{\out{<celltype name>}}}
}
\value{
dataframe (N x 2; N = number of query samples): Blood and tissue-resident immune cell identification result is stored in "origin_tb". Row names are query sample IDs.
}
\description{
Perform blood and tissue-resident immune cell identification using Originator
}
