% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Plot.Bar.doCellTypeProportion}
\alias{Plot.Bar.doCellTypeProportion}
\title{Plot$Bar$doCellTypeProportion}
\arguments{
\item{config}{Dataset config}

\item{data}{Seurat object}
}
\value{
List of plots
}
\description{
Generate barplots for cell type proportion of both blood and tissue resident
cells, w/ and w/o cell types that are not of interests
}
