% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Plot.Bar}
\alias{Plot.Bar}
\title{Plot$Bar}
\description{
An object with functions that help generate bar plots
}
\section{Functions}{

\itemize{
\item \code{\link[=Plot.Bar.doCellTypeProportion]{Plot$Bar$doCellTypeProportion}}
}
}

