% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Plot.Heatmap.buildCellTypeProportion}
\alias{Plot.Heatmap.buildCellTypeProportion}
\title{Plot$Heatmap$buildCellTypeProportion}
\arguments{
\item{config}{List of dataset configs}
}
\value{
2 cell type percentage dataframe, 1 for tissue and 1 for blood
}
\description{
Build cell type proportion dataframes for a list of datasets. Tissue-blood
annotated data must be stored as \verb{data/$\{config$name\}_tb_annotated.rds}.
}
