% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Plot.Heatmap.doProportion}
\alias{Plot.Heatmap.doProportion}
\title{Plot$Heatmap$doProportion}
\arguments{
\item{df}{Proportion dataframe, dataset by cell type}

\item{title}{Plot title, default empty string}

\item{label}{If true, label heatmap cells with values. Default true}
}
\value{
Plot object
}
\description{
Generate percentage proportion heatmap plots
}
