% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Plot.Heatmap}
\alias{Plot.Heatmap}
\title{Plot$Heatmap}
\description{
An object with functions that help generate heatmap plots
}
\section{Functions}{

\itemize{
\item \code{\link[=Plot.Heatmap.buildTissueBloodProportion]{Plot$Heatmap$buildTissueBloodProportion}}
\item \code{\link[=Plot.Heatmap.buildCellTypeProportion]{Plot$Heatmap$buildCellTypeProportion}}
\item \code{\link[=Plot.Heatmap.doProportion]{Plot$Heatmap$doProportion}}
}
}

