% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\docType{data}
\name{Plot}
\alias{Plot}
\title{Plot}
\format{
An object of class \code{environment} of length 3.
}
\usage{
Plot
}
\description{
An object with functions that help generate plots
}
\section{Objects}{

\itemize{
\item \code{\link[=Plot.Umap]{Plot$Umap}}
\item \code{\link[=Plot.Bar]{Plot$Bar}}
\item \code{\link[=Plot.Heatmap]{Plot$Heatmap}}
}
}

\keyword{datasets}
