% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Plot.Umap.doTissueBlood}
\alias{Plot.Umap.doTissueBlood}
\title{Plot$Umap$doTissueBlood}
\arguments{
\item{config}{Dataset config}

\item{data}{Seurat object}
}
\value{
List of plots
}
\description{
Generate UMAP plots, w/ and w/o labels, w/ and w/o tissue blood annotation,
w/ and w/o cell types that are not of interests
}
