% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Plot.Umap}
\alias{Plot.Umap}
\title{Plot$Umap}
\description{
An object with functions that help generate UMAP plots
}
\section{Functions}{

\itemize{
\item \code{\link[=Plot.Umap.run]{Plot$Umap$run}}
\item \code{\link[=Plot.Umap.runIfNotExists]{Plot$Umap$runIfNotExists}}
\item \code{\link[=Plot.Umap.doTissueBlood]{Plot$Umap$doTissueBlood}}
}
}

