% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Plot.Umap.runIfNotExists}
\alias{Plot.Umap.runIfNotExists}
\title{Plot$Umap$runIfNotExists}
\arguments{
\item{data}{Seurat object}
}
\value{
UMAP calculated data
}
\description{
Run UMAP on a Seurat object, if not already run
}
