% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{Utils.calcCellTypeProportion}
\alias{Utils.calcCellTypeProportion}
\title{Utils$calcCellTypeProportion}
\arguments{
\item{data}{Seurat object}

\item{ann_col}{Annotation column name}
}
\value{
Percentage dataframe with 3 columns "cellType", "proportion" and "x".
"x" column values are all same string "proportions"
}
\description{
Calculate cell type proportion for simple bar plots
}
