% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{Utils.clearMetaFeatures}
\alias{Utils.clearMetaFeatures}
\title{Utils$clearMetaFeatures}
\arguments{
\item{data}{Seurat object}
}
\value{
Updated Seurat object
}
\description{
Clear meta.features for a Seurat object. Run this before integration step.

See \url{https://github.com/satijalab/seurat/issues/2317#issuecomment-555577298}
}
