% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{Utils.dropNaCellTypeLevels}
\alias{Utils.dropNaCellTypeLevels}
\title{Utils$dropNaCellTypeLevels}
\arguments{
\item{config}{Dataset config}

\item{data}{Seurat object}
}
\value{
Updated Seurat object
}
\description{
Drop cell types from data that have zero observations
}
