% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\docType{data}
\name{Utils}
\alias{Utils}
\title{Utils}
\format{
An object of class \code{environment} of length 5.
}
\usage{
Utils
}
\description{
An object with miscellaneous utils
}
\section{Functions}{

\itemize{
\item \code{\link[=Utils.dropNaCellTypeLevels]{Utils$dropNaCellTypeLevels}}
\item \code{\link[=Utils.subsetCellTypesInConfig]{Utils$subsetCellTypesInConfig}}
\item \code{\link[=Utils.calcCellTypeProportion]{Utils$calcCellTypeProportion}}
\item \code{\link[=Utils.renameGeneNames]{Utils$renameGeneNames}}
\item \code{\link[=Utils.clearMetaFeatures]{Utils$clearMetaFeatures}}
}
}

\keyword{datasets}
