% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{Utils.renameGeneNames}
\alias{Utils.renameGeneNames}
\title{Utils$renameGeneNames}
\arguments{
\item{data}{Seurat object}

\item{newnames}{List of new gene names}
}
\value{
Updated Seurat object
}
\description{
Rename gene names for a Seurat object

Modified from \url{https://github.com/satijalab/seurat/issues/1049#issuecomment-726734005}
}
