% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{Utils.subsetCellTypesInConfig}
\alias{Utils.subsetCellTypesInConfig}
\title{Utils$subsetCellTypesInConfig}
\arguments{
\item{config}{Dataset config}

\item{data}{Seurat object}
}
\value{
Updated Seurat object
}
\description{
Subset data to keep cell types of interests
}
