% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Values.R
\docType{data}
\name{Values}
\alias{Values}
\title{Values}
\format{
An object of class \code{environment} of length 2.
}
\usage{
Values
}
\description{
An object with const values
}
\section{Vals}{

\itemize{
\item \code{\link[=Values.common_immune_cells]{Values$common_immune_cells}}
\item \code{\link[=Values.wholeblood_immune_cell_type_map]{Values$wholeblood_immune_cell_type_map}}
}
}

\keyword{datasets}
