% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Values.R
\name{Values.wholeblood_immune_cell_type_map}
\alias{Values.wholeblood_immune_cell_type_map}
\title{Values$wholeblood_immune_cell_type_map}
\description{
Map of cell type names from common immune cells to the corresponding
annotation in the wholeblood data
}
